Rename test files to snake case
