+++ cypress/integration/banners/slack_notification_banner.ts
