+++ cypress/integration/patch/patch_restart_modal.ts
